Separe the methods of cummunicate with API of budget
