+++ app/controllers/budget_controller.rb
@@ -4,15 +4,16 @@ class BudgetController < ApplicationController
   		find_public_agency
         @list_expense_month = get_list_expenses_by_period(@public_agency.id, "Janeiro", 2015, "Dezembro", 2015)
         @expense_find = 1
-      begin
+      #begin
         @list_budget_month = subtract_expenses_on_budget(@public_agency.id, 2015)
-      rescue Exception => error
+      #rescue Exception => error
+        puts "aqui"
         flash[:error] = error
         @list_budget_month = []
         @expense_find = 0
-      end
-        @list_expense_month.to_json
-        @list_budget_month.to_json
+      #end
+        @list_expense_month = @list_expense_month.to_json
+        @list_budget_month = @list_budget_month.to_json
   	end
 
   def get_list_expenses_by_period(id_public_agency,first_month="Janeiro",first_year=0000,last_month="Dezembro",last_year=9999)
@@ -128,11 +129,12 @@ class BudgetController < ApplicationController
 	def subtract_expenses_on_budget(id_public_agency,year)
 		expense = get_list_expenses_by_period(id_public_agency, "Janeiro", year, "Dezembro", year)
 		budget_array = []
-    begin
-      budgets = get_budget(id_public_agency, year)
-  		budget_array = create_budget_array(expense ,budgets, year)
-    rescue
-      raise "Não foi possível obter o orçamento do ano #{year} do Órgão Público desejado"
+	    begin
+	      budgets = BudgetAPI.get_budget(id_public_agency, year)
+	    #  puts budgets
+	  		budget_array = create_budget_array(expense ,budgets, year)
+	    rescue
+	      raise "Não foi possível obter o orçamento do ano #{year} do Órgão Público desejado"
 		end
     return budget_array
 	end
@@ -140,135 +142,16 @@ class BudgetController < ApplicationController
   def create_budget_array(expenses, budgets, year)
 		budget_array = []
 		budget = budgets[0]
-		if budget['year'] == year
+		if (budget['year']+1) == year
 			for i in 0..11	
 				value = expenses[i][1]
 				budget["value"] -= value
 				budget_array << budget['value']
 			end
 		end
-	
+		#puts budget_array
 		return budget_array
 	end
 
-  def get_budget(public_agency_id, year = 'Todos')
-    data_api = obtain_api_data(public_agency_id,year)
-    data_budget = parse_json_to_hash(data_api)
-    valid_response = valid_data?(data_budget)
-    budget_years = []
-    if valid_response
-      add_budget_array(budget_years, data_budget)
-    else
-      raise "Não foi possível obter o valor da API do orçamento"
-    end
-    return budget_years
-  end
 
-  def add_budget_array(budget_years,data_budget)
-    budgets = data_budget['results']['bindings']
-    budgets.each do |budget|
-      budget_by_year = create_budget_year(budget)
-      budget_years << budget_by_year
-    end
-  end
-  
-  def create_budget_year(budget)
-    year = budget["ano"]["value"]
-    value_budget = budget["somaProjetoLei"]["value"]
-    budget_by_year = {}
-    if !year.nil? && !value_budget.nil?
-      budget_by_year = {'year' => year.to_i,'value'=>value_budget.to_i}
-    else
-      # One of informations is null, can't add to the array
-    end
-    return budget_by_year
-  end
-
-  def obtain_api_data(public_agency_id,year)
-    data_api = ""
-    begin
-      url_query = get_url(public_agency_id,year )
-      #puts "#{url_query}"
-      uri_query = URI.parse(url_query)
-      data_api = Net::HTTP.get(uri_query)
-    rescue
-        raise "Não foi possível conectar a API"
-    end
-    return data_api
-  end
-
-  def valid_data?(budget_hash)
-    valid_data = true
-    # puts budget_hash
-    results_hash = budget_hash['results']
-    if !results_hash.nil? && !results_hash.empty?
-      bindings_array = results_hash['bindings']
-      if !bindings_array.nil? && !bindings_array.empty?
-        # Data in the budget_hash is valid
-      else
-        valid_data = false
-      end
-    else
-      valid_data = false
-    end
-    return valid_data
-  end
-
-  def parse_json_to_hash(data_api)
-    budget_year = {}
-    begin
-      budget_year = JSON.parse(data_api)
-    rescue
-      raise 'Não foi possivel conventer os dados da API do orçamento'
-    end
-    return budget_year
-  end
-
-  def get_url( public_agency_id, year = 'Todos')
-    begin_url = 'http://aondebrasil.com:8890/sparql?default-graph-uri=&query='
-
-    end_url = '&debug=on&timeout=&format=application%2Fsparql-results%2Bjson'\
-    '&save=display&fname='
-
-    year = year.to_s
-    public_agency_id = public_agency_id.to_s
-
-    url_query = generate_query(public_agency_id,year)
-    url = begin_url + url_query + end_url
-
-#    puts "\n\n\n\n#{url}\n\n\n"
-
-    return url
-  end
-
-  def generate_query(public_agency_id, year = 'Todos')
-    prefix = 'PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> '\
-            'PREFIX loa: <http://vocab.e.gov.br/2013/09/loa#>'
-    
-    year_query = query_for_year(year)
-
-    # puts year_query
-    query = 'SELECT ?ano, (SUM(?valorProjetoLei) AS ?somaProjetoLei) WHERE {'\
-    '?itemBlankNode loa:temExercicio ?exercicioURI . '+year_query+
-    '?exercicioURI loa:identificador ?ano . '\
-    '?itemBlankNode loa:temUnidadeOrcamentaria ?uoURI . '\
-    '?uoURI loa:codigo "'+public_agency_id+'" . '\
-    '?itemBlankNode loa:valorProjetoLei ?valorProjetoLei . }'
-    
-    # puts query
-    url_query = URI.encode(prefix+query)
-
-    return url_query
-  end
-  def query_for_year(year)
-    year = year.to_s
-    
-    year_query = ''
-    if year != 'Todos'
-      year_query = '?exercicioURI loa:identificador '+year+' . '
-    else
-      # line_query empty
-    end
-    return year_query
-  end
 end
