+++ app/controllers/function_controller.rb
@@ -53,11 +53,19 @@ class FunctionController < ApplicationController
 
 	def insert_expenses_functions(year_init,year_final,month_init,month_final,day_init,day_final)
 
-		@function = Function.new
-		expenses = @function.get_expenses_by_function(year_init,year_final,month_init,month_final,day_init,day_final)
+		
+		expenses = get_expenses_by_function(year_init,year_final,month_init,month_final,day_init,day_final)
 		expense_hash = convert_to_a_hash(expenses)
 		correct_datas = filter_datas_in_expense(expense_hash)
 		
+	end
+	def get_expenses_by_function(first_year,last_year,first_month,last_month,first_day,last_day)
+
+		start = Date.new(first_year,first_month,first_day)
+		end_of = Date.new(last_year,last_month,last_day)
+
+  		Function.joins(:expense).where("DATE(payment_date) BETWEEN ? AND ?", start, end_of).select("YEAR(payment_date) as date_test,sum(expenses.value) as sumValue,functions.description").group("YEAR(payment_date),functions.description").to_json
+
 	end
 
 	def convert_to_a_hash(expenses)
